azure_operations/pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"


[project]
name = "azure_operations"
version = "1.0.0"
description = "Copying content from one place to another in Azure"
readme = {file = "README.md", content-type = "markdown"}
requires-python = ">=3.8"
license = {file = "LICENSE.md"}
classifiers = [
    "Programming Language :: Python",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Private :: Do Not Upload"
]

dependencies = [
    "azure-core==1.30.1",
    "azure-storage-blob==12.19.1",
    "certifi==2024.2.2",
    "cffi==1.16.0",
    "charset-normalizer==3.3.2",
    "cryptography==42.0.5",
    "idna==3.7",
    "isodate==0.6.1",
    "pycparser==2.22",
    "requests==2.31.0",
    "setuptools==69.1.1",
    "six==1.16.0",
    "typing_extensions==4.11.0",
    "urllib3==2.2.1",
    "wheel==0.43.0",
    "pydantic==2.6.4",
    "pytest==8.1.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
filterwarnings = [
    "error",
    "ignore::UserWarning",
    'ignore:function ham\(\) is deprecated:DeprecationWarning',
]
